- added Classic99 targetted Windows build (although, the makefile will probably be removed. use the solution.) - made solution file work for 'classic99' target - made cgetc wait for key release before proceeding (unless buffered via kbhit) - ifdef out unfinished dsr_gpldsrlnk.c - fix f18a includes to be local - rename F18A initial palette due to windows conflicts - reorder resetF18a to reduce flash (did that a while ago and never pushed) - some minor fixes to testlib - fix 64 character mode for little endian targets
The Classic99 target lets some software be compiled for Windows and run with Classic99 as the video target. You can also send CPU reads and writes, but they are much slower (and most would be on the Windows application anyway). Interrupts are not working, and CRU is inaccessible, and keyboard input is global (as there's no window to focus). But testlib and example work.

To use it, you need Classic99 up and running. I am working on a version which shares VDP memory, which makes testlib very fast. Without that, the other mode uses UDP packets, which is available in the source but not currently enabled. I'll probably ship it soon. CPU memory can't really be shared with the current architecture, which is a pain but like I said above, there's little need to talk to CPU memory if you are running in Windows. Right now it's just used to upload an infinite loop to prevent the CPU from interrupting your work. (It uses the interrupt hook to take charge, so there is a small chance of it not writing the two bytes atomically enough and causing a crash.)

Anyway, not too useful yet. Might be an interesting precursor to a Windows target standalone though, if I just pull the VDP and sound parts of Classic99 over...